app.js: socket.io connection and join-room secured.
Previously code would allow anyone to connect to the socket.io from
clients like postman or some other custom client that supports
websocket connection. But now we are validating the token before
connecting the client to websocket.

Also we are not taking the rooms where the client wants to join
from the client, rather we are authenticating the user by the
token and from the authentic user_id we are querying the database
to get the rooms where the client is allowed to join.